requirements.txt
chromadb==0.4.24
llama_index==0.10.50
openai==1.35.7
pdf2docx==0.5.8
pdfkit==1.0.0
python-dotenv==1.0.1
streamlit==1.31.1
llama-index-embeddings-openai
llama-index-embeddings-azure-openai
llama-index-llms-azure-openai
llama-index-vector-stores-chroma
pysqlite3-binary